Rename all the Scala names
