+++ data_api/snapshot_generator/src/main/scala/uk/ac/wellcome/platform/snapshot_convertor/modules/AkkaS3ClientModule.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.snapshot_convertor.modules
+package uk.ac.wellcome.platform.snapshot_generator.modules
 
 import javax.inject.Singleton
